+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to Hourly Sensor will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.1.4] - 2026-08-02
+
+### Added
+
+- Preserve the most recently completed period's value in the `last_period`
+  attribute, even after that period leaves the rolling window.
+
 ## [v0.1.3] - 2026-08-01
 
 ### Fixed